🔨(project) bind edX sources to host directories
During development and testing, it is convenient to inspect the edX
sources. Therefore, we bind two Docker volumes containing the edX sources
(from the edx platform and the Python modules) to the host.
We also bind port 3000 from the edX LMS to facilitate connecting a remote
debugger for the LMS.